+++ R/ael01_nollt.R
@@ -6,7 +6,8 @@
 #' @param dataset (`character`) the name of a table in the `adam_db` object.
 #' @param default_formatting (`list`) the default format of the listing columns. See [`rlistings::as_listing`].
 #' @param col_formatting (`list`) the format of specific listing columns. See [`rlistings::as_listing`].
-#' @returns the main function returns an `rlistings` object.
+#' @param ... additional arguments passed to [`rlistings::as_listing`].
+#' @returns the main function returns an `rlistings` or a `rl_list` object.
 #'
 #' @details
 #'  * Removes duplicate rows.
@@ -35,14 +36,23 @@ ael01_nollt_main <- function(adam_db,
   assert_list(default_formatting, types = "fmt_config", names = "unique")
   assert_list(col_formatting, null.ok = TRUE, types = "fmt_config", names = "unique")
   assert_flag(unique_rows)
-  as_listing(
-    adam_db[[dataset]],
+
+  ret <- execute_with_args(
+    as_listing,
+    df = adam_db[[dataset]],
     key_cols = key_cols,
     disp_cols = disp_cols,
     default_formatting = default_formatting,
     col_formatting = col_formatting,
-    unique_rows = unique_rows
+    unique_rows = unique_rows,
+    ...
   )
+
+  if (is(ret, "list")) {
+    do_call(rl_list, ret)
+  } else {
+    ret
+  }
 }
 
 #' @describeIn ael01_nollt Preprocessing
@@ -72,7 +82,11 @@ ael01_nollt_pre <- function(adam_db,
 #' @returns the postprocessing function returns an `rlistings` object or an `ElementaryTable` (null report).
 #'
 ael01_nollt_post <- function(tlg, ...) {
-  if (nrow(tlg) == 0) tlg <- null_report
+  if (is(tlg, "rl_list")) {
+    if (length(tlg) == 0) tlg <- null_report
+  } else {
+    if (nrow(tlg) == 0) tlg <- null_report
+  }
 
   tlg
 }