+++ tests/testthat/test-ael01_nollt.R
@@ -47,3 +47,18 @@ test_that("ael01_nollt can handle some missing values", {
   res <- expect_silent(run(ael01_nollt, proc_data))
   expect_snapshot(cat(export_as_txt(res, lpp = 100)))
 })
+
+test_that("ael01_nollt listing can be split by an additional variable", {
+  res <- expect_silent(
+    run(
+      ael01_nollt,
+      syn_data,
+      dataset = "admh",
+      key_cols = c("MHBODSYS", "MHDECOD"),
+      disp_cols = "MHTERM",
+      split_into_pages_by_var = "SEX"
+    )
+  )
+  expect_list(res, type = "listing_df")
+  expect_snapshot(cat(export_as_txt(res, lpp = 100)))
+})